Moves DataGrid specific code to dataGridColumns
