+++ libs/orchestrator-ui-components/src/lib/components/Table/index.ts
@@ -1,2 +1,6 @@
 export * from './Table';
+export * from './TableHeaderCell';
 export * from './columns';
+
+export * from './DataGridTable';
+export * from './dataGridColumns';
